+++ mythril/contract/util.py
