--- mythril/rpc/constants.py
@@ -1,8 +1,0 @@
-BLOCK_TAG_EARLIEST = 'earliest'
-BLOCK_TAG_LATEST   = 'latest'
-BLOCK_TAG_PENDING  = 'pending'
-BLOCK_TAGS = (
-    BLOCK_TAG_EARLIEST,
-    BLOCK_TAG_LATEST,
-    BLOCK_TAG_PENDING,
-)
