+++ tests/native_test.py
@@ -1,5 +1,5 @@
 import json
-from mythril.ether.soliditycontract import SolidityContract
+from mythril.contract.soliditycontract import SolidityContract
 
 from mythril.laser.ethereum.state import GlobalState, MachineState
 from mythril.laser.ethereum import svm
